Change "Integer" to "BigInteger" for chat_id

+++ statistics_bot/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, Boolean, Float, String
+from sqlalchemy import Column, Integer, Boolean, Float, String, BigInteger
 from .database import Base
 
 # Create model for database
@@ -6,7 +6,7 @@ class User(Base):
     __tablename__ = 'user'
 
     _id = Column(Integer, primary_key=True)
-    chat_id = Column(Integer)
+    chat_id = Column(BigInteger)
     username = Column(String)
     plot_mode = Column(Boolean, default=True)
     digits = Column(Integer, default=6)
